Cleaning architecture for client integration
