+++ Cargo.toml
@@ -10,13 +10,12 @@ crate-type = ["cdylib"]
 
 [dependencies]
 flume = "0.11.0"
-futures-util = "0.3.28"
 num_cpus = "1.16.0"
 once_cell = "1.18.0"
 smol = "1.3.0"
-tokio = { version = "1.28.1", features = ["rt", "rt-multi-thread", "macros"] }
-wtransport = "0.1.4"
-serde = { version = "1", features = ["derive"] }
+tokio = { version = "1.32.0", features = ["rt", "rt-multi-thread", "macros"] }
+# wtransport = "0.1.4"
+wtransport = { git = "https://github.com/BiagioFesta/wtransport", branch = "master" }
 base64 = "0.21.4"
 rcgen = "0.11.1"
 ring = "0.16.20"
