+++ mod/deno.ts
@@ -33,47 +33,52 @@ const sender = new Deno.UnsafeCallback(
 
 sender.ref();
 let serverPTR;
+let new_connection;
 const decoder = new TextDecoder();
 try {
   const certpath = encodeBuffPtr("./certs/cert.pem");
   const keypath = encodeBuffPtr("./certs/key.pem");
 
-  serverPTR = lib.symbols.proc_init( sender.pointer, 4433 ,true, 0, 100, certpath[0], certpath[1], keypath[0], keypath[1]);
-  const new_connection = new Deno.UnsafeCallback(
+  serverPTR = lib.symbols.proc_server_init( sender.pointer, 4433 ,true, 0, 100, certpath[0], certpath[1], keypath[0], keypath[1]);
+  new_connection = new Deno.UnsafeCallback(
 	{
 	  parameters: [ "pointer" ],
 	  result: "void",
 	},
 	async (client) => {
 		console.log("DENO : New connection");
-		// const DBuffer = new Uint8Array(65536);
 		const DBufferB = new Uint8Array(65536);
-		//fill the buffer with 0
-		DBufferB.fill(0);
-		lib.symbols.proc_init_client_streams(serverPTR!, client, DBufferB, DBufferB.byteLength);
+
+		lib.symbols.proc_server_init_streams(client, DBufferB, DBufferB.byteLength);
 		const StreamBuffer = new ReadableStream({
 
 			async pull(controller) {
-				const nread = await lib.symbols.proc_recv_datagram(serverPTR!, client);
-				if (nread > 0) {
-					controller.enqueue(DBufferB.slice(0, nread as number));
+				try {
+					const nread = await lib.symbols.proc_recv_datagram(client);
+					if (nread > 0) {
+						controller.enqueue(DBufferB.subarray(0, nread as number));
+					}
+				} catch (e) {
+					controller.error(e);
 				}
 			},
-			cancel(_) {
+			cancel() {
 				console.error("[Error] Stream cancelled");
 			}
 		});
 		for await (const chunk of StreamBuffer) {
-			const msg = parseInt(decoder.decode(chunk));
-			console.log(msg);
+			// const msg = parseInt(decoder.decode(chunk));
+			// console.log(decoder.decode(chunk));
+			lib.symbols.proc_send_datagram(client, chunk, chunk.length);
 		}
 	},
   );
   new_connection.ref(); 
-  lib.symbols.proc_listen(serverPTR, new_connection.pointer);
+  lib.symbols.proc_server_listen(serverPTR, new_connection.pointer);
   console.info("Server is running");
 } catch (e) {
 	sender.unref();
+	new_connection?.unref();
   	console.error(e);
 }
 if (!serverPTR) throw new Error("Server is not running");
