+++ src/certificate.rs
@@ -27,7 +27,7 @@ pub struct SelfCertificate {
 }
 
 /// Generates a self-signed certificate for WebTransport connections.
-pub fn generate_certificate<S: AsRef<str>>(common_name: S) -> Result<SelfCertificate> {
+pub fn generate_certificate<S: AsRef<str>>(common_name: S, start: OffsetDateTime, end: OffsetDateTime) -> Result<SelfCertificate> {
     let keypair = KeyPair::generate(&PKCS_ECDSA_P256_SHA256)?;
     let digest = digest(&SHA256, &keypair.public_key_der());
     let fingerprint = Base64Engine.encode(digest);
@@ -39,19 +39,14 @@ pub fn generate_certificate<S: AsRef<str>>(common_name: S) -> Result<SelfCertifi
     cert_params.distinguished_name = dname;
     cert_params.alg = &PKCS_ECDSA_P256_SHA256;
     cert_params.key_pair = Some(keypair);
-    cert_params.not_before = OffsetDateTime::now_utc()
-        .checked_sub(Duration::days(2))
-        .unwrap();
-    cert_params.not_after = OffsetDateTime::now_utc()
-        .checked_add(Duration::days(2))
-        .unwrap();
+    cert_params.not_before = start;
+    cert_params.not_after = end;
 	cert_params.extended_key_usages = vec![ExtendedKeyUsagePurpose::ServerAuth, ExtendedKeyUsagePurpose::ClientAuth];
     cert_params.key_usages = vec![
 		rcgen::KeyUsagePurpose::DigitalSignature,
 		rcgen::KeyUsagePurpose::KeyCertSign,
 	];
 	cert_params.is_ca = IsCa::Ca(BasicConstraints::Unconstrained);
-	cert_params.subject_alt_names = vec![rcgen::SanType::DnsName(common_name.as_ref().to_string())];
 
     let certificate = rcgen::Certificate::from_params(cert_params)?;
 
@@ -62,3 +57,15 @@ pub fn generate_certificate<S: AsRef<str>>(common_name: S) -> Result<SelfCertifi
     })
 }
 
+
+#[no_mangle]
+pub extern "C" fn proc_gencert(buffpath: *mut u8) -> usize {
+    //get the underlying buffer and use it to return the path to the cert
+    let path = unsafe { std::ffi::CStr::from_ptr(buffpath as *const i8) };
+    let path = path.to_str().unwrap();
+	let start = OffsetDateTime::now_utc().checked_add(Duration::days(2)).unwrap();
+    let cert = generate_certificate("localhost", start, start).unwrap();
+    std::fs::write(format!("{}/cert.pem", path), cert.certificate).unwrap();
+    std::fs::write(format!("{}/key.pem", path), cert.key).unwrap();
+    path.len()
+}
